refactor: 공용 도메인 모듈 확장, User 서비스 계층 feature 단위 분리, GraphQL 루트 정리
