+++ src/features/conversation/conversation.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+
+import { ConversationRepository } from './repositories/conversation.repository';
+import { ConversationDomainService } from './services/conversation-domain.service';
+
+@Module({
+  providers: [ConversationRepository, ConversationDomainService],
+  exports: [ConversationRepository, ConversationDomainService],
+})
+export class ConversationModule {}
