+++ src/features/conversation/index.ts
@@ -1,0 +1,3 @@
+export { ConversationModule } from './conversation.module';
+export { ConversationRepository } from './repositories/conversation.repository';
+export { ConversationDomainService } from './services/conversation-domain.service';
