+++ src/features/core/root.graphql
@@ -1,0 +1,14 @@
+"""
+ISO 8601 DateTime 스칼라
+"""
+scalar DateTime
+
+type Query {
+  """스키마 루트 유지를 위한 no-op 필드"""
+  _queryNoop: Boolean
+}
+
+type Mutation {
+  """스키마 루트 유지를 위한 no-op 필드"""
+  _noop: Boolean
+}
