+++ src/features/order/index.ts
@@ -1,0 +1,4 @@
+export { OrderModule } from './order.module';
+export { OrderRepository } from './repositories/order.repository';
+export { OrderDomainService } from './services/order-domain.service';
+export { OrderStatusTransitionPolicy } from './policies/order-status-transition.policy';
