+++ src/features/order/order.types.graphql
@@ -1,0 +1,8 @@
+"""공용 주문 상태 타입(향후 user/seller 공용으로 사용)"""
+enum OrderStatusType {
+  SUBMITTED
+  CONFIRMED
+  MADE
+  PICKED_UP
+  CANCELED
+}
