+++ src/features/product/index.ts
@@ -1,0 +1,3 @@
+export { ProductModule } from './product.module';
+export { ProductRepository } from './repositories/product.repository';
+export { ProductDomainService } from './services/product-domain.service';
