+++ src/features/product/services/product-domain.service.ts
@@ -1,0 +1,4 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class ProductDomainService {}
